Rename hunk range "line end" to "line count".
The `Range` class represents a change as an inclusive range of lines: `[start, end]`.  However, per [1], unified format represents a hunk range as the starting line number of the change and the count of lines in the change (which can possibly be zero).  In fact, `UnifiedDiffParser` stores the line counts it reads from the stream as the "line end" in each `Range` it constructs.  Thus, the usage of `Range` internally is inconsistent with its API.  This is also potentially confusing for a library user who may choose to construct a `Diff` programmatically (e.g. for unit testing) and is relying on the Javadocs to describe the model behavior.

This change simply renames all occurrences of "line end" associated with `Range` to "line count" for consistency.

[1] https://en.wikipedia.org/wiki/Diff_utility#Unified_format